requirements.txt
pyTelegramBotApi
Pillow